requirements.txt
# Custom ETL Engine – Python Dependencies
# =========================================
# Install with: pip install -r requirements.txt

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# YAML configuration
PyYAML>=6.0

# Web dashboard
Flask>=3.0.0

# Testing (built-in unittest is used, but pytest is more powerful)
pytest>=7.0.0
pytest-cov>=4.0.0
